// web/src/lib/api/generated/models/githubComMantonxViewraInternalApplicationMediaMediaResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * ViewRA Media Server API
 * Self-hosted media server for movies, TV shows, and music
 * OpenAPI spec version: 0.0.1
 */

export interface GithubComMantonxViewraInternalApplicationMediaMediaResponse {
  /** aac, mp3, etc. */
  audio_codec?: string
  /** bits per second */
  bitrate?: number
  /** mkv, mp4, etc. */
  container_format?: string
  created_at?: string
  duration?: number
  file_path?: string
  file_size?: number
  /** frames per second */
  frame_rate?: number
  /** Video height in pixels */
  height?: number
  id?: number
  /** True for trailers, deleted scenes, etc. */
  is_extra?: boolean
  library_id?: number
  title?: string
  /** 'movie', 'tv_episode', 'music_track' */
  type?: string
  updated_at?: string
  /** h264, hevc, etc. */
  video_codec?: string
  /** Technical video metadata */
  width?: number
}