// web/src/lib/api/generated/models/githubComMantonxViewraInternalApplicationMoviesMediaVariantResponse.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * ViewRA Media Server API
 * Self-hosted media server for movies, TV shows, and music
 * OpenAPI spec version: 0.0.1
 */

export interface GithubComMantonxViewraInternalApplicationMoviesMediaVariantResponse {
  audio_codec?: string
  container_format?: string
  duration?: number
  file_path?: string
  file_size?: number
  height?: number
  id?: number
  language?: string
  video_codec?: string
  width?: number
}